INCLUDE: Added the `shell.nix` file.
This is only required for me. My system does not have a C compiler,
and this allows me to run a temporary shell with a c compiler. If
the "nix idea" seems interesting, give it a look. NixOS is the coolest
tool I have used in a pretty long time!